Apply valid size to grid input before generating
It may take some time before the grid is created. The interface should
reflect the most recent state or state transition.

Entering a grid size of 1,000 will be capped at 100. The input should
update to 100 so the user doesn't anticipate a 1,000 by 1,000 grid when
only a 100 by 100 grid is being worked on.